Delete creaturedex controller
