+++ app.py
@@ -1,6 +1,5 @@
 from flask_app.controllers import (
     writingdex_controller,
-    creaturedex_controller,
     index_controller,
     writing_controller, 
     creature_controller,
